requirements.txt
pythonnet==3.0.0a2
lxml
pywebview
EbookLib
beautifulsoup4
marko
PyGObject
pyinstaller
